pyproject.toml
[build-system]
requires = ["setuptools>=69"]
build-backend = "setuptools.build_meta"

[project]
name = "xcron"
version = "0.1.0"
description = "Per-project schedule reconciliation prototype for native OS schedulers"
readme = "SPEC.md"
requires-python = ">=3.9"
license = { text = "Proprietary" }
authors = [
  { name = "Dariusz Debowczyk" }
]
dependencies = [
  "PyYAML>=6.0,<7.0",
  "jsonschema>=4.23,<5.0",
  "structlog>=24.0,<26.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0,<9.0",
]

[project.scripts]
xcron = "apps.cli.main:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["apps*", "libs*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
